fix bug

+++ OpenglLearn/core/math/matrix.hpp
@@ -10,7 +10,7 @@
 
 #include <glm/glm.hpp>
 
-static double degrees_to_radians(double degrees) {
+static float degrees_to_radians(float degrees) {
     return degrees * M_PI / 180.0;
 }
